Further improvements to CACE questions
Made over several questions to insterstitials becasue
it make sense for all systems to inherit controls.

Renamed SAMS SYSTEM to SYSTEM LEVEL because the controls
should work for all systems, not just SAMS. Adjusted the output
accordingly.

Removed older se_cace_elk because that was a quick mockup of those
inherited controls and now we are inheriting all of the controls
from our parsing scripts.